+++ lib/dart_telegram_bot.dart
@@ -2,4 +2,3 @@ export 'src/entities/internal/bot.dart';
 export 'src/entities/internal/exceptions/api_exceptions.dart';
 export 'src/entities/internal/exceptions/lib_exceptions.dart';
 export 'src/entities/internal/helpers/bot_command_parser.dart';
-export 'src/entities/internal/helpers/enum_helper.dart';