+++ lib/src/entities/telegram/force_reply.dart
@@ -1,19 +1,19 @@
 import 'dart:convert';
 
-import 'package:dart_telegram_bot/telegram_entities.dart';
+import '../../../telegram_entities.dart';
 
 class ForceReply extends ReplyMarkup {
   bool forceReply = true;
   bool? selective;
 
-  ForceReply(
-    this.forceReply, {
+  ForceReply({
+    this.forceReply = true,
     this.selective,
   });
 
   static ForceReply fromJson(Map<String, dynamic> json) {
     return ForceReply(
-      json['force_reply'],
+      forceReply: json['force_reply'],
       selective: json['selective'],
     );
   }